rework: Move cli functions to files
This also exposes the modules runnable with the python -m command.
Running modules allows the coverage command to collect coverage data on process executions.
This will evenutally mean pytests can use pexpect to not only test the cli commands but also get coverage on them
and provide some nice example useage.
