+++ setup.cfg
@@ -58,14 +58,14 @@ all =
 
 [options.entry_points]
 console_scripts =
-    inet-nm-tmux = inet_nm.runner_cli:nm_tmux
-    inet-nm-exec = inet_nm.runner_cli:nm_exec
-    inet-nm-check = inet_nm.check:cli_count_board_inventory
-    inet-nm-commission = inet_nm.commissioner:commission
-    inet-nm-tty-from-uid = inet_nm.commissioner:cli_tty_from_uid
-    inet-nm-update-commissioned = inet_nm.commissioner:update_commissioned
-    inet-nm-free = inet_nm.locking:cli_release_all_locks
-    inet-nm-update-from-os = inet_nm.update_os_board_list:cli_update_boards_from_os
+    inet-nm-tmux = inet_nm.cli_tmux:main
+    inet-nm-exec = inet_nm.cli_exec:main
+    inet-nm-check = inet_nm.cli_check:main
+    inet-nm-commission = inet_nm.cli_commission:main
+    inet-nm-tty-from-uid = inet_nm.cli_tty_from_uid:main
+    inet-nm-update-commissioned = inet_nm.cli_update_commissioned:main
+    inet-nm-free = inet_nm.cli_free:main
+    inet-nm-update-from-os = inet_nm.cli_update_from_os:main
 
 
 [tool:pytest]
