+++ src/inet_nm/commissioner.py
@@ -1,13 +1,10 @@
 """This module commissions USB boards."""
-import argparse
 import random
-import sys
 from typing import List, Optional
 
 import pyudev
 
 import inet_nm._helpers as hlp
-import inet_nm.config as cfg
 from inet_nm._helpers import nm_prompt_choice, nm_prompt_confirm, nm_prompt_input
 from inet_nm.data_types import NmNode
 
@@ -161,107 +158,3 @@ def mock_device():
         driver="mock",
         mock=True,
     )
-
-
-def commission():
-    """CLI to commission a USB board."""
-    parser = argparse.ArgumentParser(description="Commission USB boards")
-    cfg.config_arg(parser)
-    parser.add_argument("-b", "--board", help="Name of the board to commission")
-    parser.add_argument(
-        "-n", "--no-cache", help="Ignore the cache", action="store_true"
-    )
-    parser.add_argument(
-        "-m",
-        "--mock-dev",
-        help="Mock a device, useful for boards that do not have USB interfaces",
-        action="store_true",
-    )
-    parser.add_argument(
-        "-i",
-        "--ignore",
-        help="Device will always be ignore, used for ttys that are connected "
-        "but not part of the inet-nm ecosystem.",
-        action="store_true",
-    )
-
-    args = parser.parse_args()
-    nodes_cfg = cfg.NodesConfig(args.config)
-    bi_cfg = cfg.BoardInfoConfig(args.config)
-    # Check early since we will need to write eventually
-    nodes_cfg.check_file(writable=True)
-    bi_cfg.check_file(writable=False)
-
-    saved_nodes = nodes_cfg.load()
-    nm_nodes = (
-        get_devices_from_tty() if args.no_cache else get_devices_from_tty(saved_nodes)
-    )
-    if args.mock_dev:
-        nm_nodes.append(mock_device())
-    print(f"Found {len(saved_nodes)} saved nodes in {args.config}")
-
-    try:
-        selected_node = select_available_node(nm_nodes)
-    except ValueError:
-        print("No available nodes found")
-        sys.exit(1)
-    if args.ignore:
-        selected_node.ignore = True
-    else:
-        binfo = bi_cfg.load()
-        selected_node.board = args.board or select_board(
-            list(binfo.keys()), selected_node
-        )
-        if selected_node.board in binfo:
-            selected_node.features_provided = binfo[selected_node.board]
-
-    nodes = add_node_to_nodes(saved_nodes, selected_node)
-    nodes_cfg.save(nodes)
-    print(f"Updated {nodes_cfg.file_path}")
-
-
-def update_commissioned():
-    """Go through the list of commissioned nodes and update features."""
-    parser = argparse.ArgumentParser(description="Update commissioned features")
-    cfg.config_arg(parser)
-    args = parser.parse_args()
-
-    nodes_cfg = cfg.NodesConfig(args.config)
-    bi_cfg = cfg.BoardInfoConfig(args.config)
-
-    nodes_cfg.check_file(writable=True)
-    bi_cfg.check_file(writable=False)
-
-    binfo = bi_cfg.load()
-    nodes = nodes_cfg.load()
-    for node in nodes:
-        if node.board in binfo:
-            node.features_provided = binfo[node.board]
-    nodes_cfg.save(nodes)
-    print(f"Updated {nodes_cfg.file_path}")
-
-
-def cli_tty_from_uid():
-    """CLI to get the TTY device node string for a given NmNode UID."""
-    parser = argparse.ArgumentParser(
-        description="Get the TTY device node string for a given NmNode UID."
-    )
-    cfg.config_arg(parser)
-    parser.add_argument("uid", help="Node UID string.")
-    args = parser.parse_args()
-
-    nodes_cfg = cfg.NodesConfig(args.config)
-    nodes_cfg.check_file(writable=False)
-
-    nodes = nodes_cfg.load()
-    try:
-        node = next(node for node in nodes if node.uid == args.uid)
-    except StopIteration:
-        print(f"Could not find node with UID {args.uid}")
-        return
-    try:
-        print(get_tty_from_nm_node(node))
-    except TtyNotPresent:
-        print(f"Could not find TTY device for {node.board} with SN {node.serial}")
-        sys.exit(1)
-    sys.exit(0)
